refactor character class into an abstract class
